Move content validator code to validator.go
