+++ pkg/content/content.go
@@ -18,30 +18,6 @@ var (
 	bmtPool = bmtlegacy.NewTreePool(swarm.NewHasher, swarm.Branches, bmtlegacy.PoolSize)
 )
 
-var _ swarm.ChunkValidator = (*ContentAddressValidator)(nil)
-
-// ContentAddressValidator validates that the address of a given chunk
-// is the content address of its contents.
-type ContentAddressValidator struct {
-}
-
-// NewContentAddressValidator constructs a new ContentAddressValidator
-func NewContentAddressValidator() swarm.ChunkValidator {
-	return &ContentAddressValidator{}
-}
-
-// Validate performs the validation check.
-func (v *ContentAddressValidator) Validate(ch swarm.Chunk) (valid bool) {
-	chunkData := ch.Data()
-	rch, err := contentChunkFromBytes(chunkData)
-	if err != nil {
-		return false
-	}
-
-	address := ch.Address()
-	return address.Equal(rch.Address())
-}
-
 // NewContentChunk creates a new content-addressed single-span chunk.
 // The length of the chunk data is set as the span.
 func NewContentChunk(data []byte) (swarm.Chunk, error) {
